Refactored code to break up artifacts from execution
