+++ crates/sui-replay-2/src/artifacts/manager.rs
@@ -1,8 +1,8 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::execution::{MoveCallInfo, ReplayCacheSummary};
-use anyhow::bail;
+use crate::artifacts::{MoveCallInfo, ReplayCacheSummary};
+use anyhow::{anyhow, bail, Result};
 use move_trace_format::format::{MoveTrace, MoveTraceReader};
 use std::{
     io::Write,
@@ -69,9 +69,9 @@ impl EncodingType {
 
 impl<'b> ArtifactManager<'b> {
     /// Creates a new `ArtifactManager` with the given base path and whether overrides are allowed.
-    pub fn new(base_path: &'b Path, overrides_allowed: bool) -> anyhow::Result<Self> {
+    pub fn new(base_path: &'b Path, overrides_allowed: bool) -> Result<Self> {
         std::fs::create_dir_all(base_path).map_err(|e| {
-            anyhow::anyhow!(
+            anyhow!(
                 "Failed to create base path for replay artifacts at {}: {e}",
                 base_path.display()
             )
@@ -134,7 +134,7 @@ impl ArtifactMember<'_, '_> {
 
     /// Deserialize the artifact into json. This should always succeed if the artifact exists and
     /// is well-formed.
-    pub fn get_json(&self) -> anyhow::Result<serde_json::Value> {
+    pub fn get_json(&self) -> Result<serde_json::Value> {
         let file_content = std::fs::read(&self.artifact_path)?;
         let contents = match self.artifact_type.encoding_type() {
             EncodingType::Json => file_content,
@@ -149,13 +149,13 @@ impl ArtifactMember<'_, '_> {
 
     /// Try to get the trace reader if the artifact type is a trace.
     /// If the artifact type is not `Trace` `None` is returned.
-    pub fn try_get_trace(&self) -> Option<anyhow::Result<MoveTraceReader<std::fs::File>>> {
+    pub fn try_get_trace(&self) -> Option<Result<MoveTraceReader<std::fs::File>>> {
         if self.artifact_type == Artifact::Trace {
             Some(
                 std::fs::File::open(&self.artifact_path)
                     .and_then(MoveTraceReader::new)
                     .map_err(|e| {
-                        anyhow::anyhow!(
+                        anyhow!(
                             "Failed to open trace file {}: {e}",
                             self.artifact_path.display()
                         )
@@ -168,14 +168,14 @@ impl ArtifactMember<'_, '_> {
 
     /// Try to get the transaction effects if the artifact type is `TransactionEffects`.
     /// If the artifact type is not `TransactionEffects` `None` is returned.
-    pub fn try_get_transaction_effects(&self) -> Option<anyhow::Result<TransactionEffects>> {
+    pub fn try_get_transaction_effects(&self) -> Option<Result<TransactionEffects>> {
         if matches!(
             self.artifact_type,
             Artifact::TransactionEffects | Artifact::ForkedTransactionEffects
         ) {
             Some(self.get_json().and_then(|json| {
                 serde_json::from_value::<TransactionEffects>(json).map_err(|e| {
-                    anyhow::anyhow!(
+                    anyhow!(
                         "Failed to deserialize transaction effects from {}: {e}",
                         self.artifact_path.display()
                     )
@@ -188,11 +188,11 @@ impl ArtifactMember<'_, '_> {
 
     /// Try to get the GasUsageReport if the artifact type is `TransactionGasReport`.
     /// If the artifact type is not `TransactionGasReport` `None` is returned.
-    pub fn try_get_gas_report(&self) -> Option<anyhow::Result<GasUsageReport>> {
+    pub fn try_get_gas_report(&self) -> Option<Result<GasUsageReport>> {
         if self.artifact_type == Artifact::TransactionGasReport {
             Some(self.get_json().and_then(|json| {
                 serde_json::from_value::<GasUsageReport>(json).map_err(|e| {
-                    anyhow::anyhow!(
+                    anyhow!(
                         "Failed to deserialize gas usage report from {}: {e}",
                         self.artifact_path.display()
                     )
@@ -205,11 +205,11 @@ impl ArtifactMember<'_, '_> {
 
     /// Try to get the ReplayCacheSummary if the artifact type is `ReplayCacheSummary`.
     /// If the artifact type is not `ReplayCacheSummary` `None` is returned.
-    pub fn try_get_cache_summary(&self) -> Option<anyhow::Result<ReplayCacheSummary>> {
+    pub fn try_get_cache_summary(&self) -> Option<Result<ReplayCacheSummary>> {
         if self.artifact_type == Artifact::ReplayCacheSummary {
             Some(self.get_json().and_then(|json| {
                 serde_json::from_value::<ReplayCacheSummary>(json).map_err(|e| {
-                    anyhow::anyhow!(
+                    anyhow!(
                         "Failed to deserialize replay cache summary from {}: {e}",
                         self.artifact_path.display()
                     )
@@ -222,11 +222,11 @@ impl ArtifactMember<'_, '_> {
 
     /// Try to get the MoveCallInfo if the artifact type is `MoveCallInfo`.
     /// If the artifact type is not `MoveCallInfo` `None` is returned.
-    pub fn try_get_move_call_info(&self) -> Option<anyhow::Result<MoveCallInfo>> {
+    pub fn try_get_move_call_info(&self) -> Option<Result<MoveCallInfo>> {
         if self.artifact_type == Artifact::MoveCallInfo {
             Some(self.get_json().and_then(|json| {
                 serde_json::from_value::<MoveCallInfo>(json).map_err(|e| {
-                    anyhow::anyhow!(
+                    anyhow!(
                         "Failed to deserialize move call info from {}: {e}",
                         self.artifact_path.display()
                     )
@@ -240,13 +240,13 @@ impl ArtifactMember<'_, '_> {
 
 /// Serialization methods for `ArtifactManager`.
 impl ArtifactMember<'_, '_> {
-    pub fn serialize_move_trace(&self, trace: MoveTrace) -> Option<anyhow::Result<()>> {
+    pub fn serialize_move_trace(&self, trace: MoveTrace) -> Option<Result<()>> {
         if self.artifact_type != Artifact::Trace {
             return None;
         }
 
         if !self.manager.overrides_allowed && self.artifact_path.exists() {
-            return Some(Err(anyhow::anyhow!(
+            return Some(Err(anyhow!(
                 "Trace file already exists at {}",
                 self.artifact_path.display()
             )));
@@ -254,7 +254,7 @@ impl ArtifactMember<'_, '_> {
 
         Some(
             std::fs::write(&self.artifact_path, trace.into_compressed_json_bytes()).map_err(|e| {
-                anyhow::anyhow!(
+                anyhow!(
                     "Failed to write trace to {}: {e}",
                     self.artifact_path.display()
                 )
@@ -262,20 +262,20 @@ impl ArtifactMember<'_, '_> {
         )
     }
 
-    pub fn serialize_artifact(&self, data: &impl serde::Serialize) -> Option<anyhow::Result<()>> {
+    pub fn serialize_artifact(&self, data: &impl serde::Serialize) -> Option<Result<()>> {
         if self.artifact_type == Artifact::Trace {
             return None;
         }
 
         if !self.manager.overrides_allowed && self.artifact_path.exists() {
-            return Some(Err(anyhow::anyhow!(
+            return Some(Err(anyhow!(
                 "File for {} already exists at {}",
                 self.artifact_type.as_str(),
                 self.artifact_path.display()
             )));
         }
         let mut file = match std::fs::File::create(&self.artifact_path).map_err(|e| {
-            anyhow::anyhow!(
+            anyhow!(
                 "Failed to create file {}: {e}",
                 self.artifact_path.display()
             )
@@ -285,21 +285,22 @@ impl ArtifactMember<'_, '_> {
         };
 
         Some(match self.artifact_type.encoding_type() {
-            EncodingType::Json => serde_json::to_writer(file, &data)
-                .map_err(|e| anyhow::anyhow!("Failed to write JSON: {e}")),
+            EncodingType::Json => {
+                serde_json::to_writer(file, &data).map_err(|e| anyhow!("Failed to write JSON: {e}"))
+            }
             EncodingType::JsonCompressed => {
                 let mut buf = Vec::new();
                 let mut compressed_buf = Vec::new();
                 if let Err(e) = serde_json::to_writer(&mut buf, &data)
-                    .map_err(|e| anyhow::anyhow!("Failed to write JSON: {e}"))
+                    .map_err(|e| anyhow!("Failed to write JSON: {e}"))
                 {
                     return Some(Err(e));
                 }
                 zstd::bulk::compress_to_buffer(&buf, &mut compressed_buf, 0)
-                    .map_err(|e| anyhow::anyhow!("Failed to compress JSON: {e}"))
+                    .map_err(|e| anyhow!("Failed to compress JSON: {e}"))
                     .and_then(|_| {
                         file.write_all(&buf).map_err(|e| {
-                            anyhow::anyhow!(
+                            anyhow!(
                                 "Failed to write compressed JSON to {}: {e}",
                                 self.artifact_path.display()
                             )
@@ -309,7 +310,7 @@ impl ArtifactMember<'_, '_> {
         })
     }
 
-    pub fn try_remove_artifact(&self) -> anyhow::Result<()> {
+    pub fn try_remove_artifact(&self) -> Result<()> {
         if self.manager.overrides_allowed {
             if let Err(e) = std::fs::remove_file(&self.artifact_path) {
                 if e.kind() != std::io::ErrorKind::NotFound {
