+++ crates/sui-replay-2/src/replay_txn.rs
@@ -10,16 +10,16 @@
 //! `get_input_objects_for_replay` is used by the `execution.rs` module but could be moved
 //! in this module and saved in the `ReplayTransaction` instance.
 
-use crate::summary_metrics::tx_metrics_reset;
 use crate::{
-    artifacts::{Artifact, ArtifactManager},
-    execution::{execute_transaction_to_effects, MoveCallInfo, ReplayCacheSummary, ReplayExecutor},
+    artifacts::{Artifact, ArtifactManager, MoveCallInfo, ReplayCacheSummary},
+    execution::{execute_transaction_to_effects, ReplayExecutor},
     replay_interface::{
         EpochStore, ObjectKey, ObjectStore, ReadDataStore, TransactionStore, VersionQuery,
     },
+    summary_metrics::tx_metrics_reset,
     tracing::save_trace_output,
 };
-use anyhow::{anyhow, bail, Context};
+use anyhow::{anyhow, bail, Context, Error, Result};
 use move_trace_format::format::MoveTraceBuilder;
 use std::collections::{btree_map::Entry, BTreeMap, BTreeSet};
 use std::time::Instant;
@@ -68,7 +68,7 @@ pub(crate) async fn replay_transaction<S: ReadDataStore>(
     data_store: &S,
     network: String,
     trace: bool,
-) -> anyhow::Result<u128> {
+) -> Result<u128> {
     let _span = info_span!("replay_tx", tx_digest = %tx_digest).entered();
     // load a `ReplayTransaction`
     tx_metrics_reset();
@@ -178,7 +178,7 @@ fn verify_txn_and_save_effects(
     artifact_manager: &ArtifactManager<'_>,
     expected_effects: &TransactionEffects,
     effects: &TransactionEffects,
-) -> anyhow::Result<()> {
+) -> Result<()> {
     // If replayed effects are different from the expected ones
     // (obtained from the chain), save the forked effects and the expected effects
     // so that they can be diffed in the output.
@@ -221,7 +221,7 @@ impl ReplayTransaction {
         txn_store: &dyn TransactionStore,
         epoch_store: &dyn EpochStore,
         object_store: &dyn ObjectStore,
-    ) -> Result<Self, anyhow::Error> {
+    ) -> Result<Self, Error> {
         debug!(op = "load_tx", phase = "start", tx_digest = %tx_digest, "load transaction");
 
         let digest = tx_digest
@@ -232,9 +232,7 @@ impl ReplayTransaction {
         // load transaction data and effects
         let transaction_info = txn_store
             .transaction_data_and_effects(tx_digest)?
-            .ok_or_else(|| {
-                anyhow::anyhow!(format!("Transaction not found for digest: {}", tx_digest))
-            })?;
+            .ok_or_else(|| anyhow!(format!("Transaction not found for digest: {}", tx_digest)))?;
         let txn_data = transaction_info.data;
         let effects = transaction_info.effects;
         let checkpoint = transaction_info.checkpoint;
@@ -249,11 +247,11 @@ impl ReplayTransaction {
         let protocol_config = match epoch_store.protocol_config(epoch) {
             Ok(Some(pc)) => pc,
             Ok(None) => {
-                tracing::error!("Protocol config missing for epoch {}", epoch);
+                error!("Protocol config missing for epoch {}", epoch);
                 return Err(anyhow!("Protocol config missing for epoch {}", epoch));
             }
             Err(e) => {
-                tracing::error!("Failed to get protocol config for epoch {}: {:?}", epoch, e);
+                error!("Failed to get protocol config for epoch {}: {:?}", epoch, e);
                 return Err(e);
             }
         };
@@ -300,7 +298,7 @@ fn load_transaction_objects(
     effects: &TransactionEffects,
     checkpoint: u64,
     object_store: &dyn ObjectStore,
-) -> Result<BTreeMap<ObjectID, BTreeMap<ObjectVersion, Object>>, anyhow::Error> {
+) -> Result<BTreeMap<ObjectID, BTreeMap<ObjectVersion, Object>>, Error> {
     // collect all package ids required by the transaction
     let mut packages = get_packages(txn_data)?;
 
@@ -332,7 +330,7 @@ fn load_transaction_objects(
 // For move calls is the package of the call.
 // For vector commands the packages of the type parameter.
 // For publish and upgrade commands, the packages of the dependencies.
-fn get_packages(txn_data: &TransactionData) -> Result<BTreeSet<ObjectID>, anyhow::Error> {
+fn get_packages(txn_data: &TransactionData) -> Result<BTreeSet<ObjectID>, Error> {
     let mut packages = BTreeSet::new();
     if let TransactionKind::ProgrammableTransaction(ptb) = txn_data.kind() {
         for cmd in &ptb.commands {
@@ -382,7 +380,7 @@ fn load_objects(
         BTreeMap<ObjectID, BTreeMap<ObjectVersion, Object>>, // objets loaded
         BTreeSet<ObjectID>,                                  // packages referenced
     ),
-    anyhow::Error,
+    Error,
 > {
     let mut packages = BTreeSet::new();
     let mut object_cache: BTreeMap<ObjectID, BTreeMap<ObjectVersion, Object>> = BTreeMap::new();
@@ -416,7 +414,7 @@ fn load_packages(
     packages: &BTreeSet<ObjectID>,
     checkpoint: u64,
     object_store: &dyn ObjectStore,
-) -> Result<BTreeMap<ObjectID, BTreeMap<ObjectVersion, Object>>, anyhow::Error> {
+) -> Result<BTreeMap<ObjectID, BTreeMap<ObjectVersion, Object>>, Error> {
     let pkg_object_keys = packages
         .iter()
         .map(|pkg_id| ObjectKey {
@@ -439,7 +437,7 @@ fn load_packages(
 fn get_txn_object_keys(
     txn_data: &TransactionData,
     effects: &TransactionEffects,
-) -> Result<Vec<ObjectKey>, anyhow::Error> {
+) -> Result<Vec<ObjectKey>, Error> {
     let input_object_ids = get_input_ids(txn_data)?;
     trace!("Input Object IDs: {:#?}", input_object_ids);
     let effects_object_ids = get_effects_ids(effects)?;
@@ -469,7 +467,7 @@ fn get_txn_object_keys(
 // That includes:
 // - the gas coins
 // -- all `CallArg::Object` to PTBs
-fn get_input_ids(txn_data: &TransactionData) -> Result<BTreeSet<ObjectKey>, anyhow::Error> {
+fn get_input_ids(txn_data: &TransactionData) -> Result<BTreeSet<ObjectKey>, Error> {
     // grab all coins
     let mut object_keys: BTreeSet<ObjectKey> = txn_data
         .gas_data()
@@ -511,7 +509,7 @@ fn get_input_ids(txn_data: &TransactionData) -> Result<BTreeSet<ObjectKey>, anyh
 }
 
 // Get the input shared objects and unchanged consensus objects from the transaction effects
-fn get_effects_ids(effects: &TransactionEffects) -> Result<BTreeSet<ObjectKey>, anyhow::Error> {
+fn get_effects_ids(effects: &TransactionEffects) -> Result<BTreeSet<ObjectKey>, Error> {
     let mut object_keys = effects
         .input_consensus_objects()
         .iter()
@@ -560,7 +558,7 @@ pub fn get_input_objects_for_replay(
     txn: &TransactionData,
     tx_digest: &TransactionDigest,
     object_cache: &BTreeMap<ObjectID, BTreeMap<u64, Object>>, // objects used by the transaction
-) -> Result<InputObjects, anyhow::Error> {
+) -> Result<InputObjects, Error> {
     let _deleted_shared_info_map: BTreeMap<ObjectID, (TransactionDigest, SequenceNumber)> =
         BTreeMap::new();
     let mut resolved_input_objs = vec![];
@@ -585,7 +583,7 @@ pub fn get_input_objects_for_replay(
                             object: ObjectReadResultKind::Object(pkg.clone()),
                         })
                     })
-                    .ok_or_else(|| anyhow::anyhow!(
+                    .ok_or_else(|| anyhow!(
                         format!(
                             "Package {} not found in transaction cache. Should have been loaded already",
                             pkg_id,
@@ -595,14 +593,14 @@ pub fn get_input_objects_for_replay(
             InputObjectKind::ImmOrOwnedMoveObject((obj_id, version, _digest)) => {
                 let object = object_cache
                     .get(obj_id)
-                    .ok_or_else(|| anyhow::anyhow!(
+                    .ok_or_else(|| anyhow!(
                         format!(
                             "Object id {}[{}] not found in transaction cache. Should have been loaded already",
                             obj_id, version,
                         )
                     ))?
                     .get(&version.value())
-                    .ok_or_else(|| anyhow::anyhow!(
+                    .ok_or_else(|| anyhow!(
                         format!(
                             "Object version {}[{}] not found in transaction cache. Should have been loaded already",
                             obj_id, version,
@@ -628,7 +626,7 @@ pub fn get_input_objects_for_replay(
                 let versions =
                     object_cache
                         .get(id)
-                        .ok_or_else(|| anyhow::anyhow!(
+                        .ok_or_else(|| anyhow!(
                             format!(
                                 "Shared Object id {} not found in transaction cache. Should have been loaded already",
                                 id,
